major refactoring: continue separating colocate context to dedicated pages
