+++ client/src/components/Login.js
@@ -1,26 +1,11 @@
-import { useGlobalContext } from '../context/GlobalContext';
-import { useAuth } from '../hooks/useAuth';
 import InputField from './InputField';
 import SigninBtn from './SigninBtn';
 import Form from './Form';
+import { useAuthContext } from '../context/AuthContext';
 
 const Login = () => {
-  const { AUTH_SIGNIN, isTabLeft, userCredential, setUserCredential } =
-    useGlobalContext();
-  const { signInUser } = useAuth();
+  const { userCredential, loginNow, checkSigninInput } = useAuthContext();
   const { email, password } = userCredential;
-
-  const checkLoginInput = (e) => {
-    setUserCredential({
-      ...userCredential,
-      [e.target.name]: e.target.value,
-    });
-  };
-
-  const loginNow = (e) => {
-    e.preventDefault();
-    signInUser(AUTH_SIGNIN, isTabLeft);
-  };
 
   return (
     <>
@@ -35,7 +20,7 @@ const Login = () => {
             name='email'
             id='email'
             value={email}
-            onChange={checkLoginInput}
+            onChange={checkSigninInput}
             required
           />
           <InputField
@@ -44,7 +29,7 @@ const Login = () => {
             name='password'
             id='password'
             value={password}
-            onChange={checkLoginInput}
+            onChange={checkSigninInput}
             required
           />
           <SigninBtn type='submit'>Einloggen</SigninBtn>
