+++ client/src/styles/_userbar.scss
@@ -1,4 +1,5 @@
 .user-bar {
+  display: inline-flex;
   z-index: 2;
   border: none;
   background-color: $bckgrnd-clr;
